+++ engine/docs/changelog.md
@@ -1,4 +1,7 @@
 # TBD
+### Features
+* Added clean endpoint which gets rid of the enclaves and the dependant containers
+
 
 # 1.7.4
 ### Changes
@@ -19,9 +22,6 @@
 * Upgrade to Kurt Core 1.36.7 to support the `@grpc/grpc-js` upgrade
 
 # 1.7.0
-### Features
-* Added clean endpoint which gets rid of the enclaves and the dependant containers
-
 ### Features
 * Added a unit test that will remind Kurtosis developers to add a breaking change to the changelog whenever they bump to a Kurt Core version that has an API break (since the engine server's API breaks when the Core API breaks due to the engine server's `KurtosisContext` returning a Core `EnclaveContext`)